Remove api-interface module and put all funtionality in api
